feat: Enable login audio playback on user interaction
Directly initiate audio playback of the login background music when a user interacts with the page (e.g., clicking to enter the front page). This circumvents browser autoplay restrictions by associating audio playback with a user gesture. If the audio source needs to be updated, it is set before playback.